refactor(header): header 컴포넌트 위치를 component에서 layout으로 이동
layout 하위에서만 사용되는 컴포넌트이므로 공용 컴포넌트단에서 레이아웃단으로 이동
