+++ src/index.ts
@@ -15,7 +15,6 @@ export * from './components/Spinner'
 export * from './components/OutlineItem'
 export * from './components/ListItem'
 export * from './components/Overlay'
-export * from './components/Header'
 export * from './components/Tooltip'
 export * from './components/Toast'
 export * from './components/SectionLabel'
@@ -40,6 +39,7 @@ export * from './layout/GNB'
 export * from './layout/GlobalHeader'
 export * from './layout/Client'
 export * from './layout/Main'
+export * from './layout/Header'
 export * from './layout/Navigations'
 export * from './layout/Side/SidePanelContent'
 export * from './layout/Side/SideViewContent'
