+++ src/layout/Navigations/Navigations.stories.tsx
@@ -9,7 +9,7 @@ import { getTitle } from '../../utils/storyUtils'
 import { styled } from '../../foundation'
 import { Icon } from '../../components/Icon'
 import { ListItem } from '../../components/ListItem'
-import { Header } from '../../components/Header'
+import { Header } from '../Header'
 import { LayoutProvider } from '../LayoutProvider'
 import { NavigationContent } from './NavigationContent'
 import Navigations from './Navigations'
