+++ src/layout/stories/LayoutPlayground/LayoutPlayGround.stories.tsx
@@ -9,12 +9,12 @@ import useSidePanelHandler from '../../../hooks/useSidePanelHandler'
 import { getTitle } from '../../../utils/storyUtils'
 import { styled, Typography } from '../../../foundation'
 import { Icon, IconSize } from '../../../components/Icon'
-import { Header } from '../../../components/Header'
 import { ListItem } from '../../../components/ListItem'
 import { LayoutProvider } from '../../LayoutProvider'
 import Client from '../../Client/Client'
 import { Main } from '../../Main'
 import GNB from '../../GNB/GNB'
+import { Header } from '../../Header'
 import Navigations from '../../Navigations/Navigations'
 import { SidePanelContent } from '../../Side/SidePanelContent'
 import { SideViewContent } from '../../Side/SideViewContent'
